Bug fixes 1

+++ index.php
@@ -95,8 +95,12 @@
     if (isset($_POST['download'])) {
 
       $p = $_POST['path'];
-      $tt->download($p);
-      echo '<script type="text/javascript">alert("Data downloaded successfully");</script>';
+      if (file_exists($p . "\\table_data_download")) {
+        echo '<script type="text/javascript">alert("This path already has folder table_data_download pls select another path or delete that folder");</script>';
+      } else {
+        $tt->download($p);
+        echo '<script type="text/javascript">alert("Data downloaded successfully");</script>';
+      }
     }
   }
 
@@ -225,7 +229,7 @@
     <form action="index.php" method="post">
       <label>
         Enter Download Path
-        <input type="text" name="path">
+        <input type="text" name="path" required>
       </label>
       <button type="submit" name="download">Download</button>
     </form>
